Fixes for repository definitions. There were many instances of OUR artifactory server repository definitions that were incorrect, not specifying exactly what sort of artifacts the repository contains. Because the artifactory server has separate repositories for release/snapshot artifacts (best practice), you MUST explicitly say in the POM what is in that repository.
If you do not do this, you will get occasions where Maven tries to download a SNAPSHOT from a release repository, causing a hard 409 'Conflict' error from Artifactory (because it is internally trying to deploy a SNAPSHOT into a repository that does not contain them).

The only was to avoid this is to be very explicit in the repository definitions. The syntax is also not quite intuitive.

For release repositories it must look something like this :-

<repository>
	<id>public-releases</id>
	<name>artifactory.javassh.com-releases</name>
	<url>http://artifactory.javassh.com/public-releases</url>
	<snapshots>
		<enabled>false</enabled>
	</snapshots>
	<releases/>
</repository>

And for snapshot repositories :-

<repository>
	<id>public-snapshots</id>
	<name>artifactory.javassh.com-snapshots</name>
	<url>http://artifactory.javassh.com/public-snapshots</url>
	<releases>
		<enabled>false</enabled>
	</releases>
	<snapshots/>
</repository>