Realm Deletion
There were many many places where when a realm was deleted, this deletion is not cascaded down to dependent resources. This is most likely to happen when the resource is in a dependent extension, and so the relationship is only specified in one direction.

To combat this, the @OnDelete annotation is used. This adds a foreign key relationship to the table that triggers a cascade when the owner is deleted. This foreign key IS added for new tables, however it will not be automatically upgraded.

So lots of upgrade scripts have been added to add these keys.

Note, it is likely there are STILL problems in this area. I have tried to catch as many as I can, but the lesser used extensions could well still have issues.