More flaws found in realm deletion. Hibernate was re-adding different foreign keys (non cascading delete) on the next start up, because the names differed from those that would be autogenerated. MySQL will accept this schema change so you end up with multiple foreign keys with different actions.
So after a few false starts, I have settled on the folowing strategu.

  * UpdateService now supports ".pre.sql" scripts. These are run BEFORE Hibnerate is initialized
  * core_2_2_1.pre.sql drops ALL foreign keys before Hibernate starts
  * Hibernate starts up and adds all the correctly cascaded keys
  * The special case is the audit table which drops it's own keys and recreates.

All of the existing SQL scripts have been removed, this is now (nearly) entirely done through the annotations and class structure.

One less than pleasant part of this was the need to move the Realm property in RealmResource, and RemoteUser/RemoteGroup principal property in RemoteUserAttributes/RemoteGroupAttributes down into the concrete implementations.
This is because this doesn't seem to work with these tables, and the @AttributeOverride annotation doesnt work either. Both of these seem to be due to known bugs in Hibernate.